mv_ddr: use ap base address from interface struct instead of define
Reviewed-on: http://vgitil04.il.marvell.com:8080/50006
Tested-by: iSoC Platform CI <ykjenk@marvell.com>

+++ ddr_topology_def.h
@@ -193,7 +193,7 @@ struct mv_ddr_iface_ops {
 
 struct mv_ddr_iface {
 	/* base addr of ap ddr interface belongs to */
-	unsigned long long ap_base;
+	unsigned int ap_base;
 
 	/* ddr interface id */
 	unsigned int id;
